+++ src/main/java/response/BlackJackResultResponse.java
@@ -1,4 +1,4 @@
-package dto;
+package response;
 
 import java.util.List;
