+++ src/main/java/response/ParticipantResponse.java
@@ -1,4 +1,4 @@
-package dto;
+package response;
 
 import java.util.List;
 import object.participant.Participant;
